Create README.md
서툰이야기 (PoorStory) README.txt
Yoondesign is a professional design group dedicated to the exploration and creation of relationships between font designs, brands, and contents, based on extensive experience and research.
There is a shyly written my story. Whenever I see you, my heart is beating. Poorstory font is designed palpitating heart as very cute and charming type design. 
It is emotional handwriting font. It has a feeling that it is a handwriting typography making different its slope degrees of initial consonant and also nurturing uneven handwriting.